requirements-dev.txt
flask
psycopg2-binary
Flask-SQLAlchemy
flask-login
bcrypt